fix the test that tested for broken caching
